modify table id type bigint

+++ DDL/articles.sql
@@ -1,8 +1,8 @@
 CREATE TABLE `articles` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` bigint NOT NULL AUTO_INCREMENT,
   `title` varchar(100) NOT NULL,
-  `user_id` int NOT NULL,
-  `shop_id` int,
+  `user_id` bigint NOT NULL,
+  `shop_id` bigint,
   `body` text,
   `status` varchar(16) NOT NULL,
   `create_at` datetime NOT NULL,
